pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = "My FastAPI website"
authors = ["Your Name <your.email@example.com>"]
#readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.43"
alembic = "^1.16.5"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"

[build-system]
requires = ["poetry-core"]
build-back = "poetry-core"